requirements.txt
feedparser
jinja2
requests
python-dateutil
beautifulsoup4
readability-lxml
